chore: add abstract class for azure resource generating bicep
